Rename array Action class
